+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   >= 1.9.3-p0 because 1.9.2 segfaults: http://redmine.ruby-lang.org/issues/3921
 * Always cleanup pry-remote DRb server and trace function when a program
   ends. Fixes [#1](https://github.com/nixme/pry-nav/issues/1).
+* **step** and **next** now check for a local file context. Prevents errors and
+  infinite loops when called from outside `binding.pry`, e.g. `rails console`.
+
 
 ## 0.0.2 (2011-11-30)