+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## UNRELEASED
+## 0.0.3 (2011-12-01)
 
 * Performance improvement: Don't trace while in the Pry console. Only works in
   >= 1.9.3-p0 because 1.9.2 segfaults: http://redmine.ruby-lang.org/issues/3921